+++ packages/examples/sdk-frontend-react/src/app/ChatUITest/ChatViewComponent.tsx
@@ -1,0 +1,30 @@
+import styled from 'styled-components';
+
+import { Section } from '../components/StyledComponents';
+
+import { ChatViewComponent } from '@pushprotocol/uiweb';
+
+
+const ChatViewComponentTest = () => {
+
+
+
+  return (
+    <div>
+      <h2>Chat UI Test page</h2>
+
+      {/* <Loader show={isLoading} /> */}
+    <ChatViewComponentCard>
+        
+      <ChatViewComponent chatId='0xCD0DAdAb45bAF9a06ce1279D1342EcC3F44845af' limit={10}/>
+      </ChatViewComponentCard>
+    </div>
+  );
+};
+
+export default ChatViewComponentTest;
+
+
+const ChatViewComponentCard = styled(Section)`
+height:60vh;
+`;
