+++ packages/examples/sdk-frontend-react/src/app/ChatUITest/ChatViewListTest.tsx
@@ -3,12 +3,12 @@ import styled from 'styled-components';
 import * as PUSHAPI from '@pushprotocol/restapi';
 import { Link } from 'react-router-dom';
 import { Section } from '../components/StyledComponents';
-import { MessageList } from '@pushprotocol/uiweb';
+import { ChatViewList } from '@pushprotocol/uiweb';
 import { EnvContext, Web3Context } from '../context';
 import { usePushChatSocket } from '@pushprotocol/uiweb';
-import { TypeBar } from '@pushprotocol/uiweb';
+import { MessageInput } from '@pushprotocol/uiweb';
 
-const MessageListTest = () => {
+const ChatViewListTest = () => {
   const { account, pgpPrivateKey } = useContext<any>(Web3Context)
 
   const { env } = useContext<any>(EnvContext);
@@ -21,19 +21,19 @@ const MessageListTest = () => {
 
       {/* <Loader show={isLoading} /> */}
 
-      <MessageListCard    >
-        <MessageList chatId='0xe19c4b204a76db09697ea54c9182eba2195542aD' limit={10} />
+      <ChatViewListCard    >
+        <ChatViewList chatId='0xe19c4b204a76db09697ea54c9182eba2195542aD' limit={10} />
 
-      </MessageListCard>
-      <TypeBar chatId='0xe19c4b204a76db09697ea54c9182eba2195542aD' isConnected={true} />
+      </ChatViewListCard>
+      <MessageInput chatId='0xe19c4b204a76db09697ea54c9182eba2195542aD' isConnected={true} />
     </div>
   );
 };
 
-export default MessageListTest;
+export default ChatViewListTest;
 
 
-const MessageListCard = styled(Section)`
+const ChatViewListCard = styled(Section)`
 height:40vh;
 background:black;
 `;
