--- packages/examples/sdk-frontend-react/src/app/ChatUITest/MessageContainer.tsx
@@ -1,32 +1,0 @@
-import { useContext, useEffect, useState } from 'react';
-import styled from 'styled-components';
-import * as PUSHAPI from '@pushprotocol/restapi';
-import { Link } from 'react-router-dom';
-import { Section } from '../components/StyledComponents';
-import { EnvContext, Web3Context } from '../context';
-import { MessageContainer } from '@pushprotocol/uiweb';
-
-
-const MessageContainerTest = () => {
-
-
-
-  return (
-    <div>
-      <h2>Chat UI Test page</h2>
-
-      {/* <Loader show={isLoading} /> */}
-    <MessageContainerCard>
-        
-      <MessageContainer chatId='24b029b8e07e60291bf9d8c0c48ff993fa1e0a99105459f7404c425c92e91bac' limit={10}/>
-      </MessageContainerCard>
-    </div>
-  );
-};
-
-export default MessageContainerTest;
-
-
-const MessageContainerCard = styled(Section)`
-height:60vh;
-`;
