+++ packages/uiweb/src/lib/components/chat/ChatViewBubble/index.ts
@@ -1,0 +1,1 @@
+export { ChatViewBubble } from './ChatViewBubble';
