+++ packages/uiweb/src/lib/components/chat/ChatViewComponent/index.ts
@@ -1,0 +1,1 @@
+export {ChatViewComponent} from './ChatViewComponent';
