+++ packages/uiweb/src/lib/components/chat/ChatViewList/index.ts
@@ -1,0 +1,1 @@
+export {ChatViewList} from './ChatViewList';
