--- packages/uiweb/src/lib/components/chat/MessageBubble/index.ts
@@ -1,1 +1,0 @@
-export { MessageBubble } from './MessageBubble';
