--- packages/uiweb/src/lib/components/chat/MessageContainer/index.ts
@@ -1,1 +1,0 @@
-export {MessageContainer} from './MessageContainer';
