+++ packages/uiweb/src/lib/components/chat/MessageInput/index.ts
@@ -1,0 +1,1 @@
+export { MessageInput } from './MessageInput';
