--- packages/uiweb/src/lib/components/chat/MessageList/index.ts
@@ -1,1 +1,0 @@
-export {MessageList} from './MessageList';
