--- packages/uiweb/src/lib/components/chat/TypeBar/index.ts
@@ -1,1 +1,0 @@
-export { TypeBar } from './TypeBar';
