+++ packages/uiweb/src/lib/components/chat/exportedTypes.ts
@@ -2,15 +2,15 @@ import type { IMessageIPFS } from '@pushprotocol/restapi';
 import { IChatTheme } from "./theme";
 import { IGroup } from '../../types'
 
-export interface IMessageListProps {
+export interface IChatViewListProps {
   chatId: string;
   limit?: number;
 }
 
-export interface IMessageContainerProps {
-  typebar?: boolean;
-  messageList?: boolean;
-  profile?: boolean; //name needs to change
+export interface IChatViewComponentProps {
+  messageInput?: boolean;
+  chatViewList?: boolean;
+  chatProfile?: boolean; //name needs to change
   chatId: string; //need confirmation on this
   limit?: number;
   emoji?: boolean;
@@ -59,7 +59,7 @@ export type GIFType = {
   height: number;
 };
 
-export interface TypeBarProps {
+export interface MessageInputProps {
   chatId: string;
   Emoji?: boolean;
   GIF?: boolean;
