+++ packages/uiweb/src/lib/icons/CaretDownGrey.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export const CalendarPurple = () => {
+export const CaretDownIcon = () => {
     return (
         <svg width="33" height="32" viewBox="0 0 33 32" fill="none" xmlns="http://www.w3.org/2000/svg">
             <path d="M25.5113 13.272L25.5114 13.2721L25.5173 13.2656C25.6713 13.0939 25.7538 12.8686 25.7499 12.6375C25.7459 12.4064 25.6558 12.1841 25.4959 12.0179C25.3356 11.8515 25.1174 11.7544 24.8867 11.7501C24.6559 11.7459 24.4344 11.835 24.2683 11.9957L24.2682 11.9956L24.2621 12.0019L16.5 20.0625L8.73789 12.0019L8.738 12.0018L8.73166 11.9957C8.56563 11.835 8.34408 11.7459 8.11332 11.7501C7.88259 11.7544 7.6644 11.8515 7.50414 12.0179C7.34416 12.1841 7.25406 12.4064 7.25013 12.6375C7.24621 12.8686 7.32869 13.0939 7.48273 13.2656L7.48262 13.2657L7.48873 13.272L15.8754 21.9823L15.8755 21.9825C16.0395 22.1526 16.2639 22.25 16.5 22.25C16.7361 22.25 16.9605 22.1526 17.1245 21.9825L17.1246 21.9823L25.5113 13.272Z" fill="#82828A" stroke="#82828A" stroke-width="0.5" />
